Refactor request.js into different modules
