+++ src/utils/request/getResponseData.js
@@ -1,0 +1,10 @@
+const getResponseData = response => {
+  const contentType = response.headers.get('content-type');
+  if (contentType && contentType.indexOf('application/json') !== -1) {
+    return response.json();
+  }
+
+  return response.text();
+};
+
+export default getResponseData;
